ai/fastapi/model_train/requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
certifi==2024.8.30
cffi==1.17.1
click==8.1.7
colorama==0.4.6
distro==1.9.0
environs==9.5.0
fastapi==0.115.4
filelock==3.16.1
fsspec==2024.10.0
grpcio==1.67.0
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
Jinja2==3.1.4
jiter==0.6.1
joblib==1.4.2
MarkupSafe==3.0.2
marshmallow==3.23.0
minio==7.2.10
mpmath==1.3.0
networkx==3.4.2
numpy==2.1.2
openai==1.52.2
packaging==24.1
pandas==2.2.3
pillow==11.0.0
protobuf==5.28.3
pycparser==2.22
pycryptodome==3.21.0
pydantic==2.9.2
pydantic_core==2.23.4
pymilvus==2.4.8
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
redis==5.2.0
scikit-learn==1.5.2
scipy==1.14.1
setuptools==75.3.0
six==1.16.0
sniffio==1.3.1
starlette==0.41.2
sympy==1.13.1
threadpoolctl==3.5.0
torch==2.5.1
torchsummary==1.5.1
torchvision==0.20.1
tqdm==4.66.6
typing_extensions==4.12.2
tzdata==2024.2
ujson==5.10.0
urllib3==2.2.3
uvicorn==0.32.0
python-multipart
requests
opencv-python>=4.9.0.80
opencv-python-headless>=4.9.0.80
